chore(rp235x): clippy clean-ups in net task + flash driver
Fix clippy nits (no behaviour change): redundant usize/u64 casts, a
deprecated elided lifetime (dhcpv4::Socket<'_>), % -> is_multiple_of in
the net task, and a pre-existing redundant u32 cast in the flash
driver's reboot status. Verified on hardware: DHCP .146, ping 12/12 0%.

+++ drv/rp235x-flash/src/main.rs
@@ -283,7 +283,7 @@ impl idl::InOrderRp235xFlashImpl for ServerImpl {
         // failure is visible at the shell instead of a silent no-op.
         userlib::hl::sleep_for(10);
         let stuck_time = wd.ctrl().read().time().bits();
-        Ok(0x8000_0000 | stuck_time as u32)
+        Ok(0x8000_0000 | stuck_time)
     }
 }
